server/requirements.txt
RPi.GPIO == 0.7.1
Flask == 2.0.3
